Remove my personal config.
The base configuration is in todo.cfg, and this is used by all tests.

Tests for actions should always use use plain mode (-p), so that the extended highlighting doesn't creep into all tests, causing a maintenance horror. There will be dedicated tests for the highlighting and filtering. It's still a good idea to keep this in the default config to check that highlighting can be disabled, filtering normally doesn't apply, and to show how it's supposed to be sequenced and configured.