browse: Limit to web resources (http:// et al.)
All other sorts of URLs are probably custom defined schemes that do not make sense to be launched through a web browser, but rather hand them to the operating system to launch them through the configured default application. For that, I'm cloning the original version of this action to a new "open" action.