XXX: Workaround for duplicate cache content under Samurize.
XXX: Capturing stderr has the strange side effect (but only when running under Samurize!) that the final type "%cacheFile%" is also captured in the cache file itself, leading to duplicated output. Couldn't find the root cause, but working around with "tee" prevents the problem, and has the slight advantage that now partial information is shown before the next full update, at the cost of an added dependency.
Show the previous cache (with a warning) when the todo.sh command yields no output at all.