colorSymbols: Rename COLOR_PROJECT to COLOR_PROJECT_EXT (same for COLOR_CONTEXT) due to overlap with built-in highlighting.
todo.txt-cli version 2.10 contains built-in highlighting of +project and @context. Because it is applied before the final filter, it makes it much more difficult to apply different line coloring in the final filters (because the previously single-colored line is now split into differing segments). Therefore, colorSymbols keeps the highlighting (so that it can be applied at the end of the final filter), but now used renamed color variables to avoid clashes with the built-in highlighting, and to keep that off by default.