Print the actually rendered tasks if TODOTXT_VERBOSE >= 1
In scheduledFilter, extract nextTask() so that we have a single place to hook into (also when filtering is skipped due to done or trashed task). Check $TODOTXT_VERBOSE and isOnRaw (so the filter is used as part of the rendermarkers action, not as the usual pre filter), and print the changed task / an empty line (to maintain the correct task numbering; fortunately, the _format() already filters those out for us) to stderr (as a side channel that does not get redirected into the todo file), which gets piped into the _format() API (so we get the usual highlighting).