BUG: math domain error in "completed projects (no open tasks)" when no such done task(s) in done.txt
When all the done tasks (for a +project or @context) are in todo.txt, but none in done.txt, the doneNum is 0, and math.log(0, 10) fails.
As the task number is only used to determine the maximum width, we can safely pass the sum of taskNum + doneNum here, and thereby avoid the problem.

Traceback (most recent call last):
  File "/home/inkarkat/.todo/actions/birdseye", line 218, in <module>
    main(sys.argv[1:])
  File "/home/inkarkat/.todo/actions/birdseye", line 211, in main
    run('+', item, isPrintOpen, isPrintCompleted)
  File "/home/inkarkat/.todo/actions/birdseye", line 187, in run
    printTaskGroups("completed %s (no open tasks)"% (prefixDescription, ), tasksWithNoIncompletes, [], {}, doneNum)
  File "/home/inkarkat/.todo/actions/birdseye", line 88, in printTaskGroups
    printTaskGroup(item, itemMaxLength, -1, " ", taskNum)
  File "/home/inkarkat/.todo/actions/birdseye", line 93, in printTaskGroup
    taskWidth = int(math.ceil(math.log(taskNum,10)))
ValueError: math domain error